requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
packaging==24.1
pillow==10.3.0